organize query adapter tests
